konfirmasi hapus dan keterangan script

+++ tabel1.php
@@ -47,7 +47,7 @@
 												echo"<td>$row[vol]</td>";?>												
 												    <td class="text-center">
                                                     <a href="modul.php?isi=input1&act=edit&id=<?php echo $row['idt1'] ?>"><button type="button" class="btn btn-info">Edit</button> 
-                                                    <a href="hapus.php?tabel=tabel1&id=<?php echo $row['idt1'] ?>"><button type="button" class="btn btn-danger">Hapus</button></a></td>                      
+                                                    <a href="hapus.php?tabel=tabel1&id=<?php echo $row['idt1'] ?>" onclick="return confirm('Anda Yakin?')"><button type="button" class="btn btn-danger">Hapus</button></a></td>                      
                                        <?php } ?>
                                     </tbody>
                                 </table>
